requirements.txt
dialogflow==0.6.0
python-dotenv==0.10.3
python-telegram-bot==11.1.0
requests==2.22.0
vk-api==11.4.0
